Refactor the replication execution
1. Use the task manager to manage the underlying execution/task
2. Use the pkg/scheduler to schedule the periodical job
3. Apply the new program model
4. Migration the old data into the new data model

+++ src/server/v2.0/handler/preheat.go
@@ -663,7 +663,7 @@ func convertTaskToPayload(model *task.Task) (*models.Task, error) {
 		ExecutionID:   model.ExecutionID,
 		ExtraAttrs:    model.ExtraAttrs,
 		ID:            model.ID,
-		RunCount:      int64(model.RunCount),
+		RunCount:      model.RunCount,
 		StartTime:     model.StartTime.String(),
 		Status:        model.Status,
 		StatusMessage: model.StatusMessage,
